+++ torch_geometric/datasets/__init__.py
@@ -107,6 +107,7 @@ from .infection_dataset import InfectionDataset
 from .ba2motif_dataset import BA2MotifDataset
 from .ba_multi_shapes import BAMultiShapesDataset
 from .ba_shapes import BAShapes
+from .inlut3d import InLUT3D
 
 import torch_geometric.datasets.utils
 
@@ -190,6 +191,7 @@ homo_datasets = [
     'BrcaTcga',
     'NeuroGraphDataset',
     'WebQSPDataset',
+    'InLUT3D',
 ]
 
 hetero_datasets = [